+++ src/components/season/season-settings-popover.tsx
@@ -6,7 +6,6 @@ import {
   setSeasonShowReorder,
   setSeasonShowThisWeek,
   setSeasonShowWishlist,
-  setSeasonStickyHeader,
   useUi,
 } from "@/store/ui";
 import { For, VStack } from "@chakra-ui/react";
@@ -15,7 +14,6 @@ import { Tooltip } from "../ui/tooltip";
 
 function SeasonSettingsPopover() {
   const {
-    seasonStickyHeader,
     seasonShowReorder,
     seasonShowCheckboxes,
     seasonShowCarsDropdown,
@@ -26,13 +24,6 @@ function SeasonSettingsPopover() {
   } = useUi();
 
   const settingsList = [
-    {
-      id: "hover",
-      text: "Highlight on track hover",
-      tooltip: "Highlight all cells with the same hovered track",
-      checked: seasonHighlight,
-      setChecked: setSeasonHighlight,
-    },
     {
       id: "checkboxes",
       text: "Show content checkboxes",
@@ -48,11 +39,11 @@ function SeasonSettingsPopover() {
       setChecked: setSeasonShowCarsDropdown,
     },
     {
-      id: "sticky",
-      text: "Sticky headers",
-      tooltip: "Keep column headers always visible.",
-      checked: seasonStickyHeader,
-      setChecked: setSeasonStickyHeader,
+      id: "hover",
+      text: "Highlight on track hover",
+      tooltip: "Highlight all cells with the same hovered track",
+      checked: seasonHighlight,
+      setChecked: setSeasonHighlight,
     },
     {
       id: "reorder",